[BadAtNames] recommended changes
