+++ README.md
@@ -1,0 +1,10 @@
+# ProjectTerrabyte
+
+## File Structure
+
+```
+server: all the files under here are under the server
+client: all the files are client side
+```
+
+GOLDEN RULE: Never reference something in the `client` folder from `server` and vice-versa
