+++ client/index.html
@@ -1,0 +1,19 @@
+<html> 
+	<head>
+		<title> Terrabite </title>
+		<style>
+			canvas {
+				width: 800px;
+				height: 600px;
+				border: 5px solid black;
+			}
+		</style>
+	</head>
+	<body>
+		<canvas id = "canvas"></canvas>
+ 	</body>
+    <footer>
+        <script src="https://cdn.jsdelivr.net/npm/socket.io-client@2/dist/socket.io.js"></script>
+        <script src="./static/game.js"></script>
+    </footer>
+</html>
