+++ server/README.md
@@ -1,0 +1,12 @@
+## Install
+
+```
+npm install
+```
+
+## Run
+
+```
+npm start # runs an instance
+npm start-dev # run server.js in nodemon, will auto-restart when you modify any files it touches
+```
