--- terrabite/Terrabite/Terrabite.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.852
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Terrabite", "Terrabite\Terrabite.vcxproj", "{251FF4CC-6B11-495E-AB13-A874EEE945A2}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{251FF4CC-6B11-495E-AB13-A874EEE945A2}.Debug|x64.ActiveCfg = Debug|x64
-		{251FF4CC-6B11-495E-AB13-A874EEE945A2}.Debug|x64.Build.0 = Debug|x64
-		{251FF4CC-6B11-495E-AB13-A874EEE945A2}.Debug|x86.ActiveCfg = Debug|Win32
-		{251FF4CC-6B11-495E-AB13-A874EEE945A2}.Debug|x86.Build.0 = Debug|Win32
-		{251FF4CC-6B11-495E-AB13-A874EEE945A2}.Release|x64.ActiveCfg = Release|x64
-		{251FF4CC-6B11-495E-AB13-A874EEE945A2}.Release|x64.Build.0 = Release|x64
-		{251FF4CC-6B11-495E-AB13-A874EEE945A2}.Release|x86.ActiveCfg = Release|Win32
-		{251FF4CC-6B11-495E-AB13-A874EEE945A2}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {EB662ED9-F5F2-4FC5-A548-237AFA92F4C2}
-	EndGlobalSection
-EndGlobal
