--- terrabite/index.html
@@ -1,17 +1,0 @@
-<html> 
-	<head>
-		<title> Terrabite </title>
-		<style>
-			canvas {
-				width: 800px;
-				height: 600px;
-				border: 5px solid black;
-			}
-		</style>
-		<scrpt src= "/socket.io/socket.io.js"></script>
-	</head>
-	<body>
-		<canvas id = "canvas"></canvas>
- 	</body>
-	<script src = "/static/game.js"></script>
-</html>
