--- terrabite/npm-debug.log
@@ -1,49 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ '/usr/local/bin/node',
-1 verbose cli   '/usr/local/bin/npm',
-1 verbose cli   'install',
-1 verbose cli   '-g',
-1 verbose cli   'n' ]
-2 info using npm@1.1.66
-3 info using node@v0.8.15
-4 verbose read json /usr/local/lib/package.json
-5 verbose read json /usr/local/lib/package.json
-6 verbose cache add [ 'n', null ]
-7 silly cache add name=undefined spec="n" args=["n",null]
-8 verbose parsed url { pathname: 'n', path: 'n', href: 'n' }
-9 silly lockFile d1854cae-n n
-10 verbose lock n /root/.npm/d1854cae-n.lock
-11 silly lockFile d1854cae-n n
-12 verbose addNamed [ 'n', '' ]
-13 verbose addNamed [ null, '' ]
-14 silly lockFile 6b58d7dc-n n@
-15 verbose lock n@ /root/.npm/6b58d7dc-n.lock
-16 silly addNameRange { name: 'n', range: '', hasData: false }
-17 verbose url raw n
-18 verbose url resolving [ 'https://registry.npmjs.org/', './n' ]
-19 verbose url resolved https://registry.npmjs.org/n
-20 info trying registry request attempt 1 at 23:19:18
-21 http GET https://registry.npmjs.org/n
-22 silly lockFile 6b58d7dc-n n@
-23 error Error: SSL Error: CERT_UNTRUSTED
-23 error     at ClientRequest.<anonymous> (/usr/local/lib/node_modules/npm/node_modules/request/main.js:525:26)
-23 error     at ClientRequest.g (events.js:192:14)
-23 error     at ClientRequest.EventEmitter.emit (events.js:96:17)
-23 error     at HTTPParser.parserOnIncomingClient [as onIncoming] (http.js:1462:7)
-23 error     at HTTPParser.parserOnHeadersComplete [as onHeadersComplete] (http.js:111:23)
-23 error     at CleartextStream.socketOnData [as ondata] (http.js:1367:20)
-23 error     at CleartextStream.CryptoStream._push (tls.js:526:27)
-23 error     at SecurePair.cycle (tls.js:880:20)
-23 error     at EncryptedStream.CryptoStream.write (tls.js:267:13)
-23 error     at Socket.ondata (stream.js:38:26)
-24 error If you need help, you may report this log at:
-24 error     <http://github.com/isaacs/npm/issues>
-24 error or email it to:
-24 error     <npm-@googlegroups.com>
-25 error System Linux 5.3.0-24-generic
-26 error command "/usr/local/bin/node" "/usr/local/bin/npm" "install" "-g" "n"
-27 error cwd /home/axiou_lepida/Desktop/terrabite
-28 error node -v v0.8.15
-29 error npm -v 1.1.66
-30 error code ESSL
-31 verbose exit [ 1, true ]
