--- terrabite/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "game",
-  "version": "1.0.0",
-  "description": "lets try to make a game",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "express": "^4.17.1",
-    "mkdirp": "^0.5.1",
-    "nodemon": "^2.0.2",
-    "socket.io": "^2.3.0"
-  },
-  "devDependencies": {}
-}
